Work on intermediary definition
This is difficult, as this is a thicket of traps for the unwary.  The
notion of "beneficiary" that I'm using here is weak in the sense that
intermediaries often benefit from harvesting information about
exchanges.

I'm happy with that in the sense that the benefit that they derive is
indirect.  That search engines (which unfortunately are very much
intermediaries, despite their function appearing to be a service), use
the click stream as a primary source of income doesn't make the value
they derive any more direct under this definition.

The other key aspect is that this allows for modification of protocol
units by the intermediary.  This is, again, a thicket of nuance that
probably could be the subject of a longer dissertation.  For me, the key
distinction is that the intermediary would not be in a position to
generate the protocol unit it sends without first having received a
similar unit from another entity.

It's possible that this opens a bigger discussion into messages that the
intermediary receives and does not forward when that was not intended by
the original sender.  Similarly, messages that the intermediary
generates unilaterally.  I've added a little bit about that to the
Security Considerations.

I've also expanded this to talk about the status of an entity as
intermediary is contextual.  Taking the example of the search engine
again, that is an HTTP, TLS, TCP, and IP endpoint, but at the layer that
matters to a human, it's an intermediary.  It's not OSI layers, as there
are many protocols that are partially terminated and partially
forwarded.  For instance, many protocols deliberately generate messages
that are similar to end-to-end messages, but are constructed to be
consumed or produced by intermediaries.  Still, that's the concept we
best understand here, so I'm sticking with it.